Fixed: ensure init.py shell commands run in bash
The Python standard says that subprocess commands run with shell=True will
run with the "sh" shell. On Fedora-alike distros (and macOS?) this ends up
being "bash" but some distros actually have "sh" installed too, and "sh"
doesn't have the "source" command so the virtualenv won't be activated.

So I've got a work-around now where we use "env" to find and start "bash"
for us, to make sure everything... well, to make sure that doesn't happen.