New: when cloning repos, fallback to GitHub if Phabricator doesn't work
This is intended for CI and people who don't have Phabricator accounts.
First we try to clone the repos from Phabricator, but if that doesn't work
then we'll clone them from the GitHub mirrors. GitHub is read-only, so
this may cause a little confusion... but it's easy enough to change the
remote URL if needed.