New: 'activate' command in the 'ncoda' script
This command activates the Lychee virtualenv in a new shell. The usual
"source" method wouldn't work because it only takes effect in a shell
that's currently running. (I.e., with "source" the virtualenv would be
deactivated at the end of the "nc" script or, if we started a new shell
in the "nc" script, the virtualenv wouldn't be active there).

This means we have to use "exit" to deactivate the virtualenv. We could
add our own one-line "deactivate" script to the virtualenv, but this would
interfere with people who still want to use "source" to activate the venv.

This also means we aren't able to (reliably) change PS1 for the shell in
which the virtualenv is active. People may be confused; I hope to fix this.

For T101